[fix] Use `process.env.USERPROFILE` if $HOME isn't set
Due to changes in node v0.10, `path.join` no longer accepts `undefined`
as a correct parameter which causes this module to fail on Windows,
which sets $USERPROFILE instead of $HOME.

Disclaimer: I have no actual idea if this'll make the module work on
Windows. It might if people have a full bash shell set up.